backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
bcrypt
PyJWT
motor
pymongo